Rename ASTNodePos to AST

+++ src/type_checker/types/construct/mod.rs
@@ -1,2 +1,4 @@
+use crate::type_checker::context::type_name::TypeName;
+
 // TODO look up range implemention of Python
-pub const RANGE: TypeName = TypeName { lit: String::from("range"), generics: vec![] };
+pub const RANGE: TypeName = TypeName::Single { lit: String::from("range"), generics: vec![] };
