+++ src/type_checker/types/primitive/mod.rs
@@ -1,4 +1,4 @@
 use crate::type_checker::context::type_name::TypeName;
 
-pub const INT: TypeName = TypeName { lit: String::from("int"), generics: vec![] };
-pub const BOOLEAN: TypeName = TypeName { lit: String::from("bool"), generics: vec![] };
+pub const INT: TypeName = TypeName::Single { lit: String::from("int"), generics: vec![] };
+pub const BOOLEAN: TypeName = TypeName::Single { lit: String::from("bool"), generics: vec![] };
